pyproject.toml
[project]
name = "langgraph-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.0.0",
    "langchain-openai>=0.3.35",
    "langchain[anthrophic]>=0.3.27",
    "langgraph>=0.6.10",
    "python-dotenv>=1.1.1",
    "mcp>=1.0.0",
]
